update chat window components
