+++ client/src/App.js
@@ -7,8 +7,8 @@ import "bootstrap/dist/css/bootstrap.min.css"
 
 function App() {
   return (
-    <ChatApp />
-    // <Signup/>
+    // <ChatApp />
+    <Signup />
 
     //   <>
