+++ client/src/components/Avatar/index.js
