+++ client/src/components/Contacts/index.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+// create the side bar for contacts here and add on the left or right side of the chat app div
