+++ client/src/components/InputBox/index.js
@@ -1,7 +1,5 @@
 import React, { Component } from 'react';
-import InputGroup from 'react-bootstrap/InputGroup';
-import FormControl from 'react-bootstrap/FormControl';
-import Button from 'react-bootstrap/Button';
+import { InputGroup, FormControl, Button } from 'react-bootstrap';
 
 class InputBox extends Component {
   render() {
