+++ client/src/components/NavBar/index.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import { Nav } from 'react-bootstrap';
+
+class NavBar extends Component {
+  render() {
+    return (
+      <nav>
+        <h1 className='justify-content-center'>message-app</h1>
+      </nav>
+    );
+  }
+}
+
+export default NavBar;
